Removed unnecessary "setting" table
Current values can be taken directly from gui controls. Event listeners on
checkboxes are also not needed anymore thanks to that change as they were
not doing anything useful.

Siren count and type edit boxes were hovewer performing validation and
did't update the setting table if value was invalid. This was misleading
for user in my opinion because what he sees on the screen is not what he
is really setting by pressing "Apply" button.

Currently validation only gives visual feedback on chat but doesn't do
anything behind the scenes.

This commit includes custom getNumber() method on GuiElement for getting
number value instead of string from edit boxes.

New table with default values for gui controls is introduced because the
old "setting" table was also playing this role before. It's better to have
it all in one place at the beginning of the script instead of hardcoded
inline in gui creation code.